Delete README.md
Climate change is one of the biggest threats to the humanity, and that massive changes must be made in how we live to revert it and/or reduce its impact. At the same time, there are also many skeptics, who either do not believe in the threat per se, or that it is caused by human activity, or do not see a need (or possibility) to do much about it. This issue and the debate are, of course, multi-faceted, but in this assignment we will look at the statistical evidence for global warming – a largely scientific, fact-based view on the situation. 

The attached case (Global Warming Revisited (A), UVA-QA-0808) describes a situation that fueled the debate in 2013 – an observation that the “Global Warming Stopped.” The case mentions the data obtained at the time and provides the links to two well-known sources of data: NASA and UK Met Office.